+++ src/main/java/com/saintdan/framework/component/ValidateHelper.java
@@ -1,8 +1,6 @@
 package com.saintdan.framework.component;
 
-import com.saintdan.framework.constant.ControllerConstant;
 import com.saintdan.framework.enums.ErrorType;
-import com.saintdan.framework.enums.OperationStatus;
 import com.saintdan.framework.param.BaseParam;
 import com.saintdan.framework.param.ClientParam;
 import com.saintdan.framework.po.User;
@@ -10,10 +8,10 @@ import com.saintdan.framework.service.ClientService;
 import com.saintdan.framework.tools.SpringSecurityUtils;
 import com.saintdan.framework.vo.ResultVO;
 import org.apache.commons.codec.binary.Base64;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.validation.BindingResult;
 
 /**
  * Validate helper.
@@ -32,51 +30,50 @@ public class ValidateHelper {
     /**
      * Validate current user, param and sign.
      *
+     * @param result            bind result
      * @param currentUser       current user
      * @param param             param
-     * @param sign              signature
+     * @param sign              sign
      * @param log               log
-     * @return                  result VO
+     * @return                  result vo
      * @throws Exception
      */
-    public ResultVO validate(User currentUser, BaseParam param, String sign, Log log) throws Exception {
-        if (currentUser == null) {
-            return resultHelper.infoResp(log, ErrorType.SYS0003);
+    public ResultVO validate(BindingResult result, User currentUser, BaseParam param, String sign, Log log) throws Exception {
+        if (result.hasErrors()) {
+            return resultHelper.infoResp(ErrorType.SYS0002, result.toString());
         }
-        ResultVO resultVO = validateParams(param);
-        if (resultVO != null) {
-            return resultVO;
-        }
-        return validateSign(param, sign, log);
+        return validate(currentUser, param, sign, log);
     }
 
     /**
-     * Validate the param.
+     * Validate current user and sign.
      *
-     * @param param             param
-     * @return                  result VO
+     * @param currentUser currentUser
+     * @param param       param
+     * @param sign        signature
+     * @param log         log
+     * @return result VO
+     * @throws Exception
      */
-    public ResultVO validateParams(BaseParam param) {
-        // Get incorrect params.
-        String validateContent = param.getIncorrectParams();
-        if (!StringUtils.isBlank(validateContent)) {
-            // If validate failed, return error message.
-            return new ResultVO(ErrorType.SYS0002.description(), OperationStatus.FAILURE,
-                    String.format(ControllerConstant.PARAM_BLANK, validateContent));
+    public ResultVO validate(User currentUser, BaseParam param, String sign, Log log)throws Exception {
+        //check currentUser
+        if (currentUser == null || currentUser.getId() == null) {
+            return resultHelper.infoResp(log, ErrorType.SYS0003);
         }
-        return null;
+        return validate(param, sign, log);
     }
 
     /**
      * Validate sign.
      *
-     * @param param             param
-     * @param sign              signature
-     * @param log               log
-     * @return                  result VO
+     * @param param param
+     * @param sign  signature
+     * @param log   log
+     * @return result VO
      * @throws Exception
      */
-    public ResultVO validateSign(BaseParam param, String sign, Log log) throws Exception {
+
+    public ResultVO validate(BaseParam param, String sign, Log log) throws Exception {
         // Prepare to validate signature.
         param.setSign(new String(Base64.decodeBase64(sign.getBytes())));
         // Get current clientId
@@ -85,8 +82,8 @@ public class ValidateHelper {
         if (!signHelper.signCheck(getPublicKeyByClientId(clientId), param, sign)) {
             // Return rsa signature failed information and log the exception.
             return resultHelper.infoResp(log, ErrorType.SYS0004);
-        }
-        return null;
+        } else
+            return null;
     }
 
     // --------------------------