+++ src/main/java/com/saintdan/framework/controller/ClientController.java
@@ -19,10 +19,13 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
 
 /**
  * Client's controller
@@ -74,7 +77,7 @@ public class ClientController {
         try {
             ClientParam param = new ClientParam();
             // Sign validate.
-            ResultVO resultVO = validateHelper.validateSign(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, log);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -103,7 +106,7 @@ public class ClientController {
             }
             ClientParam param = new ClientParam();
             // Sign validate.
-            ResultVO resultVO = validateHelper.validateSign(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, log);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -131,7 +134,7 @@ public class ClientController {
             }
             ClientParam param = new ClientParam(Long.valueOf(id));
             // Sign validate.
-            ResultVO resultVO = validateHelper.validateSign(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, log);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -153,12 +156,13 @@ public class ClientController {
      * @return          client's result
      */
     @RequestMapping(value = ResourceURL.CLIENTS + "/{id}" + ResourceURL.SIGN, method = RequestMethod.POST)
-    public ResultVO update(@CurrentUser User currentUser, @PathVariable String id, @PathVariable String sign, ClientParam param) {
+    public ResultVO update(@CurrentUser User currentUser, @PathVariable String id, @PathVariable String sign, @Valid ClientParam param, BindingResult result) {
         try {
             if (StringUtils.isBlank(id)) {
                 return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
             }
-            ResultVO resultVO = validateHelper.validate(currentUser, param, sign, log);
+            // Validate current user, param and sign.
+            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, log);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -187,7 +191,7 @@ public class ClientController {
             }
             ClientParam param = new ClientParam(Long.valueOf(id));
             // Sign validate.
-            ResultVO resultVO = validateHelper.validateSign(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, log);
             if (resultVO != null) {
                 return resultVO;
             }