+++ src/main/java/com/saintdan/framework/param/GroupParam.java
@@ -1,8 +1,8 @@
 package com.saintdan.framework.param;
 
 import com.saintdan.framework.annotation.SignField;
-import com.saintdan.framework.annotation.ValidationField;
 
+import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 
 /**
@@ -20,7 +20,7 @@ public class GroupParam extends BaseParam implements Serializable {
     private Long id; // role's ID.
 
     @SignField
-    @ValidationField
+    @NotNull(message = "Name cannnot be null.")
     private String name; // role's name
 
     @SignField