+++ src/main/java/com/saintdan/framework/param/UserParam.java
@@ -1,8 +1,9 @@
 package com.saintdan.framework.param;
 
 import com.saintdan.framework.annotation.SignField;
-import com.saintdan.framework.annotation.ValidationField;
 
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 import java.io.Serializable;
 
 /**
@@ -20,15 +21,17 @@ public class UserParam extends BaseParam implements Serializable {
     private Long id; // user's ID
 
     @SignField
-    @ValidationField
+    @NotNull(message = "Usr cannnot be null.")
+    @Size(min = 4, max = 50)
     private String usr; // username
 
     @SignField
-    @ValidationField
+    @NotNull(message = "Pwd cannnot be null.")
+    @Size(min = 8, max = 50)
     private String pwd; // password
 
     @SignField
-    @ValidationField
+    @NotNull(message = "Name cannnot be null.")
     private String name; // user's name
 
     @SignField